requirements.txt
annotated-types==0.7.0
anyio==4.4.0
beautifulsoup4==4.12.3
certifi==2024.7.4
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
cssselect==1.2.0
dnspython==2.6.1
fastapi==0.112.1
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.15.4
gdown==5.2.0
gensim==4.3.3
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.8.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
jieba3k==0.35.1
joblib==1.4.2
lxml==5.2.2
lxml_html_clean==0.2.0
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.26.4
pandas==2.1.4
pillow==10.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pymongo==4.8.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
requests-file==2.1.0
rfc3986==1.5.0
scipy==1.13.1
sgmllib3k==1.0.0
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
soupsieve==2.5
starlette==0.38.2
tinysegmenter==0.3
tldextract==5.1.2
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wrapt==1.16.0
